+++ docs/CHANGELOG.md
@@ -16,6 +16,9 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Support for progressive jpeg loading. ([#877])
 * New `application_root_dir()` function in `amethyst_utils`. ([#831])
 * Load node names for glTF prefabs. ([#905])
+* `IsometricTransform` abstracting away 2D isometric math. ([#895])
+* `XYCameraSystem` to control a 2D camera along the X and Y axis. ([#895])
+* `SpriteSheetFormat` to load sprite positions into a sprite sheet from a RON file. ([#895])
 
 ### Changed
 * Sprites contain their dimensions and offsets to render them with the right size and desired position. ([#829], [#830])
@@ -28,6 +31,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Derive `Debug` and `PartialEq` for `amethyst_input::Axis`. ([#903], [#904])
 
 ### Removed
+* `SpriteRenderData` and `WithSpriteRender` 2D utilities. Replaced by `SpriteRender`.
 
 ### Fixed
 * Material ids in GLTF loader caused multiple GLTF files to get incorrect materials applied ([#915])
@@ -41,6 +45,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#887]: https://github.com/amethyst/amethyst/pull/887
 [#892]: https://github.com/amethyst/amethyst/pull/892
 [#877]: https://github.com/amethyst/amethyst/pull/877
+[#895]: https://github.com/amethyst/amethyst/pull/895
 [#896]: https://github.com/amethyst/amethyst/pull/896
 [#831]: https://github.com/amethyst/amethyst/pull/831
 [#902]: https://github.com/amethyst/amethyst/pull/902